[SPARK-53340][CONNECT] Make StreamingQueryCommand in SparkConnectPlanner side effect free
